4.21

+++ osa4/controllers/blogs.js
@@ -21,7 +21,6 @@ blogsRouter.post('/', async (request, response) => {
             return response.status(401).json({ error: 'token invalid' })
         }
         const user = await User.findById(decodedToken.id)
-        console.log(user.id)
         const blog = new Blog({
             title: body.title,
             author: body.author,
@@ -37,10 +36,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: 'token invalid' })
+        }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    const blogUser = blog.user
+    console.log(blogUser, user.id==blogUser)
+    console.log(!(user.id==blogUser))
+    if (!(user.id==blogUser)){
+        return response.status(401).json({error: 'cant delete other users blog'})
+    } else {
       await Blog.findByIdAndDelete(request.params.id)
       response.status(204).end()
-   
+    }
   })
 
 blogsRouter.put('/:id', async (request, response) => {
